Refactored mapping of properties to API calls
Or rather the code that does the mappings...

Mappings are now mostly split per API call, and presented in a
declarative fashion. SpotRequest mappings are expressed in as much as
they differ from RunInstance mappings. This is meant to be both more
readable and maintainable.

Warnings will now be generated for options that are not processed,
except when creating auto-scaling groups, where mappings for
CreateLaunchConfiguration and CreateAutoScalingGroup will ignore each
others parameters, as the two calls are meant to be complementary and
used together.

Command-line action name changed from 'run-instance' to simply 'run',
trying to detach it a bit from the RunInstances AWS API call.