Make tags work in auto-scaling groups
Had to work around boto, given that it will happily propagate missing
optional arguments as 'Null' and AWS is not very happy about that.

For now, this makes all tags propagate to started instances, which
generally fits with the instance abstraction. In the future we might
want to explore group-only tagging (PropagateAtLaunch=false).