Implement file manifests
File manifests specify a list of files to be grabbed from S3 and
deployed in the instance. This is made available through the 'manifest'
keyword, which acts as a proxy for 'user-data', much like the 'puppet'
field.

The directory for puppet "manifests" was renamed to avoid confusion with
the file manifests now introduced.