Reimplement boto.utils.fetch_file
Going this route, a single connection can be shared across all
downloads. More importantly, the number of roundtrips is reduced to 1/3,
as fetch_file would do an initial GET to validate the bucket, then a
HEAD to validate the key and finally a GET to grab its contents.

This also ensures s3:GetObject is the only action that the script
requires permission for.