Notify user that ASG / LC already exist
hc2000 does not attempt to update existing auto-scaling groups or launch
configurations, silently proceeding as if it had created them.

Adding some information to the output as a first step. Going forward
this could become an error requiring explicit action to circumvent or
ASG and LC could be automatically updated to match the provided
definition as is already done when creating IAM roles.