Add scope to manifest-related error messages
An initial_scope parameter is added to the generic validate function and
Context constructor.

Scope string is now tracked when validating manifests, including
user-data entry and file name if relevant. The scope may not be the user
scope, as the entries in the 'manifest' key will be injected to
'user-data' where validation actually happens.

validate function extracted so it can be reused in instance validation.
Soon!