Don't terminate generated images on startup :-)
The original implementation was flawed in the use of a semaphore for the
termination condition, except that every clone of the generated instance
would attempt to terminate itself.

Instead, the image creation semaphore now contains the instance ID that
is generating the image. When it reboots, only the original image will
terminate itself. Clones get to live their own independent lives.

Since termination is meant to be a one way street, the termination
semaphore is dropped altogether. If termination fails, it's likely a
permission error and the semaphore doesn't change anything -- the
instance should not go into cyclic reboots.