Add 'cloud-config' plugin
As it turns out, there can only be one cloud-config file per instance
otherwise CloudInit will misbehave or, more likely, silently drop
commands. This means the earlier puppet support is broken when used
together with explicit #cloud-config entries in user-data.

With this, we make cloud-config a first-class citizen in instance
configuration, which also enables some validation and compression of the
resulting yaml. For the time being, the only validation provided is the
requirement of valid yaml.