Put resource in charge of its own plugins
While it may be nice to keep plugin code outside of the resource (and
while this is way too simple for now :-), plugins are really a way to
extend and connect different resource types.

As such, they should be in charge of their own customization points. In
the case of instance, there could potentially be pre-launch and
post-launch customization points.