Add blob resource
This is not meant to replace boto.s3. But will get you some way as far
as fetching and uploading objects goes.

Adding this here for convenience and to simplify sharing connections and
credentials with the rest of hc2000. Connection object returned by
hc2002.aws.s3 is a boto.s3.connection.S3Connection and can be used as
such.