Move instance profile creation earlier
Eventual consistency can be a pain. For instance profiles, creating them
and immediately attempting to create ASG launch configurations that
refer to them (such as could be done in a deployment script orchestrated
with hc2002 :-), can lead to surprises.

While this isn't likely to fix that problem, well, anything that may
help is welcome.

This also simplifies code in the unexpected code path, potentially at
the cost of one extra IAM call.