AMI Image Search plugin
This exposes EC2's DescribeImages API call. All filters currently
available in the API are exposed, but not validated. This construct does
not accept queries by ImageId, as that would be rather pointless here.

Wildcard characters '*' and '?' are recognized in filter values. Online
AWS docs have more info:

    http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeImages.html

This allows instance definitions to be (a bit)gq more independent of
specific regions and enables them to effectively track the latest
version of a specific distribution.`