Make role deletion robust towards interrupted creation
If role creation gets interrupted along the way, we could end up with
either a detached IAM role or a detached IAM instance-profile.

Don't let that hinder deletion, but also don't try to be any smarter: if
the names don't match, it's not an hc2002 role and the function will
fail without deleting role or instance profile.