Update block-device mappings in ASG launch configs to latest boto
We used to work-around some boto brokenness, but they have caught up to
us and offered different brokenness. On the up side, this is progressing
in a good direction, where the same types are expected for
ec2:RunInstances and autoscaling:CreateLaunchConfiguration.

The last remaining snag is that the CreateLaunchConfiguration wrapper
now expects the mapping inside an array.

I might go ahead and submit a pull request to boto to fix that and
break things for us... Again.