Ignore validation errors when deleting roles
IAM roles and instance profiles have different requirements on names:
(currently) the former is limited to 64 charaters, the latter to 128.

Given that hc2002 treats roles and instance profiles as a single
resource this leads to a leak: hc2002 will create an IAM instance
profile for a name that fails validation as an IAM role; attempting to
delete that role, however, will fail with an exception, leaving behind
the instance profile originally created.

With this change, we ensure that validation errors won't prevent
deletion of the role resource (and associated IAM instance profile, role
and policies), allowing such previously created instance profiles to be
deleted.

No new validation is added at resource creation time for now.