Map count to both min- and max-count, when calling RunInstances
Essentially making it stand in for an exact count. Previously, we would
pick up boto's default of 1 instance and map count only to the maximum
number of instances.

At the same time, made validation stricter to ensure minimum or maximum
are not specified along with exact count. Previously, count and
max-count would trump each other, with no defined rule for precedence.